+++ posthog-samples/posthog-android-sample/src/main/java/com/posthog/android/sample/NormalActivity.kt
@@ -37,6 +37,13 @@ class NormalActivity : ComponentActivity() {
 //            startActivity(Intent(this, NothingActivity::class.java))
 //            finish()
             // Check if the "enable_network_request" feature flag is enabled
+
+            try {
+                throw RuntimeException("Test error")
+            } catch (e: Throwable) {
+                PostHog.captureException(e, mapOf("am-i-stupid" to true))
+            }
+
             val isNetworkRequestEnabled = PostHog.isFeatureEnabled("enable_network_request", false)
 
             if (isNetworkRequestEnabled) {