+++ posthog/src/main/java/com/posthog/PostHogEventName.kt
@@ -8,6 +8,7 @@ public enum class PostHogEventName(public val event: String) {
     GROUP_IDENTIFY("\$groupidentify"),
     CREATE_ALIAS("\$create_alias"),
     FEATURE_FLAG_CALLED("\$feature_flag_called"),
+    EXCEPTION("\$exception"),
     ;
 
     public companion object {